requirements.txt
# TokenMachine Server Dependencies
# Updated: 2026-01-21
# Server 端：控制平面，不包含推理引擎

# ============================================================================
# Web Framework
# ============================================================================
fastapi==0.127.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.45
psycopg2-binary==2.9.9
alembic==1.18.0

# ============================================================================
# Cache
# ============================================================================
redis==5.2.1
hiredis==3.3.0

# ============================================================================
# GPU Monitoring (Server 端监控 GPU 状态)
# ============================================================================
pynvml==11.5.0

# ============================================================================
# HTTP Client
# ============================================================================
httpx==0.26.0

# ============================================================================
# Monitoring
# ============================================================================
prometheus-client==0.21.1

# ============================================================================
# Logging
# ============================================================================
loguru==0.7.3

# ============================================================================
# Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ============================================================================
# Model Hub Integration
# ============================================================================
huggingface-hub==0.35.0
modelscope>=1.0.0

# ============================================================================
# System Monitoring
# ============================================================================
psutil==5.9.8

# ============================================================================
# Async
# ============================================================================
asyncio-mqtt==0.16.1
aiofiles==23.2.1


